added processing signals and clean_up functions, more customization

+++ build_hyb_uefi.sh
@@ -34,16 +34,23 @@ boot_file_src="http://cdimage.debian.org/cdimage/unofficial/non-free/cd-includin
 boot_file="firmware-7.2.0-amd64-netinst.iso"
 boot_file_src_checksum="74a675e7ed4a31c5f95c9fc21f63a5e60cc7ed607055773ffb9605e55c4de4cb"
 boot_file_checksum_type="sha256"
-bootstrap_cfg="bootstrap/my_preseed.cfg"
+# destination bootstrap file (kickstart, preseed) - this is relative path to build_initrd_dir
+bootstrap_cfg="preseed.cfg"
+bootstrap_cfg_src="bootstrap/my_preseed.cfg"
+# orig. efi image - this is relative path to build_iso_dir
+boot_efi="boot/grub/efi.img"
+
 # name for the new image
 output_image="${boot_file_src_path}/custom-debian-7.2.0-amd64-firmware-uefi.iso"
 output_image_volid="Custom-debian-7.2.0-amd64"
-# orig. initrd file - this is relative path from BUILD_DIR
+# orig. initrd file - this is relative path to build_iso_dir
 initrd_file="install.amd/initrd.gz"
 # where build custom initrd
-INITRD_DIR="debian-initrd-build"
+build_initrd_dir="debian-initrd-build"
 # where build custom image
-BUILD_DIR="debian-iso-build"
+build_iso_dir="debian-iso-build"
+# remove build directories after build (or during error build); 1 mean yes
+clean_up_build=1
 
 # functions
 function download_iso {
@@ -69,52 +76,51 @@ function download_iso {
 }
 
 function prepare_iso_build {
-    if [[ -d ${BUILD_DIR} ]]; then
+    if [[ -d ${build_iso_dir} ]]; then
         # nothing to do
-        printf "[INFO] ${BUILD_DIR} already exists.\n"
+        printf "[INFO] ${build_iso_dir} already exists.\n"
         return 0
     fi
     # create tmp directory for iso mount
     tmp_isomount=$(TMPDIR=. mktemp -d)
-    mkdir -p "${BUILD_DIR}"
+    mkdir -p "${build_iso_dir}"
     mount -o loop "${boot_file_src_path}/${boot_file}" "${tmp_isomount}"
-    rsync -v -a -H "${tmp_isomount}/" "${BUILD_DIR}/"
+    rsync -v -a -H "${tmp_isomount}/" "${build_iso_dir}/"
     umount "${tmp_isomount}"
     rm -rf "${tmp_isomount}"
 }
 
 function prepare_initrd_build {
-    if [[ -d ${INITRD_DIR} ]]; then
+    if [[ -d ${build_initrd_dir} ]]; then
         # nothing to do
-        printf "[INFO] ${INITRD_DIR} already exists.\n"
+        printf "[INFO] ${build_initrd_dir} already exists.\n"
         return 0
     fi
     # unpack orig. initrd
-    mkdir -p "${INITRD_DIR}"
-    pushd "${INITRD_DIR}"
-        gzip -d < "../${BUILD_DIR}/${initrd_file}" | cpio --extract --verbose --make-directories --no-absolute-filenames
+    mkdir -p "${build_initrd_dir}"
+    pushd "${build_initrd_dir}"
+        gzip -d < "../${build_iso_dir}/${initrd_file}" | cpio --extract --verbose --make-directories --no-absolute-filenames
     popd
 }
 
 function deploy_custom_initrd {
     # add preseed file to initrd and pack it
-    pushd "${INITRD_DIR}"
+    pushd "${build_initrd_dir}"
         rm -f "../initrd.gz.custom"
-        cp -H "../${bootstrap_cfg}" "preseed.cfg"
+        cp -H "../${bootstrap_cfg_src}" "${bootstrap_cfg}"
         find . | cpio -H newc --create --verbose | gzip -9 > "../initrd.gz.custom"
     popd
     # deploy custom initrd
-    cp "initrd.gz.custom" "${BUILD_DIR}/${initrd_file}"
-    ls -la "${BUILD_DIR}/${initrd_file}"
+    cp "initrd.gz.custom" "${build_iso_dir}/${initrd_file}"
+    ls -la "${build_iso_dir}/${initrd_file}"
     # update md5sum.txt file - this is a good practice, not a must have
-    pushd "${BUILD_DIR}"
+    pushd "${build_iso_dir}"
         md5sum $(find -type f) > md5sum.txt
     popd
 }
 
 function build_uefi_hybrid {
     printf "[INFO] Making ${output_image} hybrid ISO image...\n"
-    #xorriso -indev "${boot_file_src_path}/${boot_file}" 2>&1 | grep "Volume id"
     xorriso -as mkisofs \
         -r -J \
         -V "${output_image_volid}" \
@@ -124,36 +130,65 @@ function build_uefi_hybrid {
         -boot-load-size 4 \
         -boot-info-table \
         -isohybrid-mbr "${boot_file_src_path}/${boot_file}" \
-        -eltorito-alt-boot -e boot/grub/efi.img \
+        -eltorito-alt-boot -e "${boot_efi}" \
         -isohybrid-gpt-basdat \
         -no-emul-boot \
         -o "${output_image}" \
-        "${BUILD_DIR}"
-    # Note: some extra/alternative options:
-    #-partition_offset 16 \
-    #-r -J -joliet-long -cache-inodes \
-    #-isohybrid-mbr "/usr/lib/syslinux/isohdpfx.bin" \
-    #-isohybrid-mbr "/usr/share/syslinux/isohdpfx.bin" \
-}
-
-function quazi_clean {
-    # ugly clean up
-    rm -rf "${BUILD_DIR}"
-    rm -rf "${INITRD_DIR}"
-    rm -rf "initrd.gz.custom"
+        "${build_iso_dir}"
 }
 
 function clean_up {
-    printf "Not implemented yet.\n"
+    if [[ ${clean_up_build} -eq 0 ]]; then
+        return 0
+    fi
+    if [[ -d "${build_iso_dir}" ]]; then
+        rm -rf "${build_iso_dir}"
+    fi
+    if [[ -d "${build_initrd_dir}" ]]; then
+        rm -rf "${build_initrd_dir}"
+    fi
+    if [[ -f "initrd.gz.custom" ]]; then
+        rm -rf "initrd.gz.custom"
+    fi
+    if [[ -d "${tmp_isomount}" ]]; then
+        rm -rf "${tmp_isomount}"
+    fi
+}
+
+function signal_clean_up {
+    printf "[INFO] Signal/Error handler - cleanup before exiting...\n"
+    clean_up
     exit 1
 }
 
 function dependencies_check {
-    printf "Not implemented yet.\n"
-    exit 1
+    local dependencies=(
+        "curl:curl"
+        "xorriso:xorriso"
+        "rsync:rsync"
+        "gzip:gzip"
+        "cpio:cpio"
+        "openssl:openssl"
+        "md5sum:coreutils"
+    )
+    for depend in "${dependencies[@]}"; do
+        local dep_cmd="${depend%%:*}"
+        local dep_name="${depend##*:}"
+        # check whether dep_name is installed - dep_cmd command exist
+        if ! command -v $dep_cmd >/dev/null 2>&1; then
+            printf "[ERROR] '${dep_cmd}' command doesn't exist - install '${dep_name}' to continue.\n"
+            exit 1
+        fi
+    done
 }
 
+
 # MAIN
+# check whether we have everything to start with script
+dependencies_check
+# signals and error handler
+trap signal_clean_up SIGHUP SIGINT SIGTERM ERR
+
 # use comments to omit some steps - e.g. if you want only update preseed file (from previous build),
 # then comment everything, but deploy_custom_initrd and build_uefi_hybrid
 download_iso
@@ -161,6 +196,4 @@ prepare_iso_build
 prepare_initrd_build
 deploy_custom_initrd
 build_uefi_hybrid
-quazi_clean
-
-#trap clean_up SIGHUP SIGINT SIGTERM ERR
+clean_up
